requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
et_xmlfile==2.0.0
eurostat==1.1.1
idna==3.10
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
tzdata==2024.2
urllib3==2.3.0
